requirements.txt
termcolor
praw
python-dotenv
